% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneQTL.R
\name{GeneQTL}
\alias{GeneQTL}
\title{This function enables to find the list of the selected genes overlapped with QTL regions.}
\usage{
GeneQTL(geneset, genelist, qtl)
}
\arguments{
\item{geneset}{vector of characters representing the names of genes/ gene ids selected from the whole gene list/space by using a gene selection method.}

\item{genelist}{N by 3 dataframe/ matrix (genes/gene ids as row names); where, N represents the number of genes in the whole gene set: first coloumn represnting the chromosomal location of genes: second coloumn representing the start position of genes in terms of basepairs: third coloumn representing the end position of genes in terms of basepairs in their respective chromosomes.}

\item{qtl}{Q by 3 dataframe/matrix (qtl names/qtl ids as row names);where, Q represents the number of qtls: first coloumn represnting the chromosomal location of qtls: second coloumn representing the start position of qtls in terms of basepairs: third coloumn representing the end position of qtls in terms of basepairs in their respective chromosomes.}
}
\description{
The function provides the list of the selected genes overlapped Quantitative Trait Loci (QTL) regions along with their genomic positions.
}
\author{
Samarendra Das
}
