% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSQMicro.R
\name{GSQMicro}
\alias{GSQMicro}
\title{This function perofrms gene set analysis with Quantitative Trait Loci for Gene expreesion data derrived from Microarray studies through integrating differential expression (DE) scores of genes.}
\usage{
GSQMicro(ExpData, class, geneLoc, size, QTLData, method)
}
\arguments{
\item{ExpData}{N by M matrix of expression values, rows as genes (N: total number of genes) and columns as samples (M: Total samples).}

\item{class}{Vector of 1 and 0 of length M shows the class labels of samples (1: case and 0: control).}

\item{geneLoc}{N by 3 dataframe/ matrix (genes/gene ids as row names); where, N represents the number of genes in the whole gene set: first coloumn represnting the chromosomal location of genes: second coloumn representing the start position of genes in terms of basepairs: third coloumn representing the end position of genes in terms of basepairs in their respective chromosomes.}

\item{size}{size of the selected gene set to be analyzed with the QTL, e.g. 100, 500, ...}

\item{QTLData}{Q by 3 dataframe/matrix (qtl names/qtl ids as row names);where, Q represents the number of qtls: first coloumn represnting the chromosomal location of qtls: second coloumn representing the start position of qtls in terms of basepairs: third coloumn representing the end position of qtls in terms of basepairs in their respective chromosomes.}

\item{method}{A character representing the Differential Expression (DE) analysis method to be used for DE analysis of expression data. It must be either T-test or F-scores or MRMR or SVM.}
}
\value{
This functions rturns the statistical QTL enrichment significance value of the gene set of size 'size' along with test statistic.
}
\description{
This function perofrms gene set analysis with Quantitative Trait Loci for Gene expreesion data derrived from Microarray studies through integrating differential expression (DE) scores of genes.
}
\author{
Samarendra Das
}
